backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
]
